refactor: 패키지 이동 및 테스트 데이터 분리
