+++ src/main/java/cart/controller/ViewController.java
@@ -1,7 +1,7 @@
 package cart.controller;
 
-import cart.domain.product.Product;
 import cart.persistance.dao.ProductDao;
+import cart.persistance.entity.product.Product;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
