+++ src/main/java/cart/persistance/entity/product/ProductImageUrl.java
@@ -1,4 +1,4 @@
-package cart.domain.product;
+package cart.persistance.entity.product;
 
 public class ProductImageUrl {
