+++ src/test/java/cart/controller/CartControllerTest.java
@@ -1,47 +1,28 @@
 package cart.controller;
 
-import cart.domain.product.Product;
+import cart.controller.dto.AddCartRequest;
 import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
 import org.apache.http.HttpStatus;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.web.server.LocalServerPort;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
-import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
-import org.springframework.jdbc.support.KeyHolder;
+import org.springframework.test.context.jdbc.Sql;
 
 import static io.restassured.RestAssured.given;
 
-
+@Sql(value = "/test.sql")
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 class CartControllerTest {
-
-    private long key;
 
     @LocalServerPort
     private int port;
 
-    @Autowired
-    private JdbcTemplate jdbcTemplate;
-
-
     @BeforeEach
     void setUp() {
         RestAssured.port = port;
-
-        final var simpleInsert = new SimpleJdbcInsert(jdbcTemplate)
-                .withTableName("product")
-                .usingGeneratedKeyColumns("id");
-        final var parameterSource = new BeanPropertySqlParameterSource(
-                Product.createWithoutId("product1", 100, "url.com"));
-        final KeyHolder keyHolder = simpleInsert.executeAndReturnKeyHolder(parameterSource);
-
-        key = keyHolder.getKey().longValue();
     }
 
     @DisplayName("GET /cart-products 요청시 존재하는 아이디 비밀번호일 시 OK 반환")
@@ -71,7 +52,7 @@ class CartControllerTest {
     void addCartProductTest() {
         given().log().all()
                 .contentType(ContentType.JSON)
-                .body(key)
+                .body(new AddCartRequest(1L))
                 .auth().preemptive().basic("user1@woowa.com", "123456")
                 .when()
                 .post("/cart-products")
@@ -82,14 +63,6 @@ class CartControllerTest {
     @DisplayName("DELETE /cart-products/{id} 요청 시 성공하면 status no content 반환")
     @Test
     void deleteCartProductTest() {
-        given()
-                .contentType(ContentType.JSON)
-                .body(key)
-                .auth().preemptive().basic("user1@woowa.com", "123456")
-                .when()
-                .post("/cart-products")
-                .then();
-
         given().log().all()
                 .auth().preemptive().basic("user1@woowa.com", "123456")
                 .when()
