+++ src/test/java/cart/dao/ProductDaoTest.java
@@ -1,21 +1,20 @@
 package cart.dao;
 
-import cart.domain.product.Product;
 import cart.persistance.dao.ProductDao;
+import cart.persistance.entity.product.Product;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.support.GeneratedKeyHolder;
-import org.springframework.jdbc.support.KeyHolder;
+import org.springframework.test.context.jdbc.Sql;
 
-import java.sql.PreparedStatement;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
+@Sql("/test.sql")
 @JdbcTest
 class ProductDaoTest {
 
@@ -54,38 +53,19 @@ class ProductDaoTest {
     @DisplayName("상품 전체를 조회한다.")
     @Test
     void shouldReturnAllProductsWhenRequest() {
-        jdbcTemplate.update("INSERT INTO product (name, price, image_url) VALUES (?, ?, ?)", "사과", 100, "domain.com");
-        jdbcTemplate.update("INSERT INTO product (name, price, image_url) VALUES (?, ?, ?)", "당근", 100, "domain.com");
 
         final List<Product> products = productDao.findAll();
 
-        assertAll(
-                () -> assertThat(products).hasSize(2),
-                () -> assertThat(products.get(0).getName()).isEqualTo("사과"),
-                () -> assertThat(products.get(0).getPrice()).isEqualTo(100),
-                () -> assertThat(products.get(0).getImageUrl()).isEqualTo("domain.com"),
-                () -> assertThat(products.get(1).getName()).isEqualTo("당근")
-        );
+        assertThat(products).hasSize(2);
     }
 
     @DisplayName("상품을 수정한다.")
     @Test
     void shouldUpdateWhenRequest() {
+
         //given
-        final KeyHolder keyHolder = new GeneratedKeyHolder();
-
-        jdbcTemplate.update(connection -> {
-            PreparedStatement preparedStatement = connection.prepareStatement(
-                    "INSERT INTO product (name, price, image_url) VALUES (?, ?, ?)", new String[]{"id"});
-            preparedStatement.setString(1, "사과");
-            preparedStatement.setLong(2, 100);
-            preparedStatement.setString(3, "domain.com");
-            return preparedStatement;
-        }, keyHolder);
-
-        long productId = keyHolder.getKey().longValue();
-        Product productToUpdate = Product.create(
-                productId,
+        final Product productToUpdate = Product.create(
+                1L,
                 "당근",
                 1000,
                 "domain.kr"
@@ -101,7 +81,7 @@ class ProductDaoTest {
                         resultSet.getString("name"),
                         resultSet.getLong("price"),
                         resultSet.getString("image_url")
-                ), productId);
+                ), 1);
 
         //then
         assertAll(
@@ -116,22 +96,7 @@ class ProductDaoTest {
     @Test
     void shouldDeleteWhenRequest() {
 
-        //given
-        final KeyHolder keyHolder = new GeneratedKeyHolder();
-
-        jdbcTemplate.update("INSERT INTO product (name, price, image_url) VALUES (?, ?, ?)", "당근", 1000, "domain.com");
-        jdbcTemplate.update(connection -> {
-            PreparedStatement preparedStatement = connection.prepareStatement(
-                    "INSERT INTO product (name, price, image_url) VALUES (?, ?, ?)", new String[]{"id"});
-            preparedStatement.setString(1, "사과");
-            preparedStatement.setLong(2, 100);
-            preparedStatement.setString(3, "domain.com");
-            return preparedStatement;
-        }, keyHolder);
-        long id = keyHolder.getKey().longValue();
-
-        //when
-        productDao.deleteById(id);
+        productDao.deleteById(1);
 
         List<Product> products = jdbcTemplate.query(
                 "SELECT id, name, price, image_url FROM product",
@@ -142,7 +107,6 @@ class ProductDaoTest {
                         resultSet.getString("image_url")
                 ));
 
-        //then
         assertThat(products).hasSize(1);
     }
 }
