+++ src/test/java/cart/domain/product/ProductNameTest.java
@@ -1,12 +1,13 @@
 package cart.domain.product;
 
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-
+import cart.persistance.entity.product.ProductName;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.EmptySource;
 import org.junit.jupiter.params.provider.ValueSource;
+
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 class ProductNameTest {
