+++ src/test/java/cart/domain/product/ProductPriceTest.java
@@ -1,9 +1,10 @@
 package cart.domain.product;
 
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-
+import cart.persistance.entity.product.ProductPrice;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 class ProductPriceTest {
